fix: fix sweeper test for datasets
This commit fixes the GQL query we issue to get a list of datasets. The
previous version resulted in an API error.

Additionally, check `managedById` before attempting to delete a dataset.
Datasets that are managed by another object should never be swept
directly.

+++ observe/observe_sweeper_test.go
@@ -97,11 +97,11 @@ func datasetSweeperFunc(s string) error {
 	for _, workspace := range workspaces {
 		result, err := client.Meta.Run(ctx, `
 		query getDatasetsInWorkspace($workspaceId: ObjectId!) {
-		    datasetsInWorkspace($workspaceId: $workspaceId) {
-			    id
+		    workspace(id: $workspaceId) {
 			    datasets {
 			        id
 			        label
+			        managedById
 			    }
 		    }
 		}`, map[string]interface{}{
@@ -112,13 +112,15 @@ func datasetSweeperFunc(s string) error {
 			return fmt.Errorf("failed to lookup datasets: %w", err)
 		}
 
-		for _, i := range result["datasetsInWorkspace"].([]interface{}) {
+		result = result["workspace"].(map[string]interface{})
+		for _, i := range result["datasets"].([]interface{}) {
 			var (
-				item  = i.(map[string]interface{})
-				label = item["label"].(string)
-				id    = item["id"].(string)
+				item        = i.(map[string]interface{})
+				label       = item["label"].(string)
+				id          = item["id"].(string)
+				managedById = item["managedById"]
 			)
-			if prefixRe.MatchString(label) {
+			if prefixRe.MatchString(label) && managedById == nil {
 				log.Printf("[WARN] Deleting %s [id=%s]\n", label, id)
 				if err := client.DeleteDataset(ctx, id); err != nil {
 					return err
